pyproject.toml
[project]
name = "fantasy-baseball-sleeper-finder"
version = "0.1.0"
description = "AI-powered dynasty auction fantasy baseball player evaluations"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Data pipeline
    "pybaseball>=2.2.7",
    "mlb-statsapi>=1.7.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # ML
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",
    "shap>=0.46.0",
    "optuna>=4.1.0",
    # LLM
    "anthropic>=0.40.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Utilities
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
